test(e2e): the live MCP walk states what it proves, its process owners are fixtures, and pair search declares at most one chain
Measured against codex-cli 0.153.2: a project-level .codex/config.toml is
never read (codex mcp list in the project answers []), only
$CODEX_HOME/config.toml is, and a throwaway CODEX_HOME cannot carry an agent
turn because it is also the credential directory (401). The spec header now
says so under a heading that names what the spec does not prove, and a new
layer 0 copies the installer's file into a throwaway CODEX_HOME and asserts
both halves through the binary itself; layer 2 keeps the -c overrides for
the agent turn.

The MCP JSON-RPC process owner moves to fixtures/studio-mcp-client.ts and the
Codex child and event owner to fixtures/codex-live-runner.ts; the spec drops
from 889 to 535 lines with one acceptance flow, discovery unchanged.

dexscreener__pairs_search declared exclusiveParamGroups over chain and
chainIds, so a query naming neither was refused at the boundary while the
parameter text, the handler summary and the transport all treat an unscoped
search as legitimate; the archived live walk shows an external agent's first
call refused for exactly that. The manifest now declares atMostOne, the
toolsnap sentence changes from "exactly one" to "at most one", and a boundary
test drives all four chain-scope shapes through validateProtocolParams.
